move relevant files to a UnifiedFrontendSolver module
