+++ src/UHF/Driver.hs
@@ -28,15 +28,15 @@ import qualified UHF.Data.SIR as SIR
 import qualified UHF.Data.SIR.PP as SIR.PP
 import qualified UHF.Diagnostic as Diagnostic
 import qualified UHF.Diagnostic.Settings as DiagnosticSettings
-import qualified UHF.Parts.TypeSolver as TypeSolver
-import qualified UHF.Parts.InfixGroup as InfixGroup
+import qualified UHF.Parts.UnifiedFrontendSolver.TypeSolver as TypeSolver
+import qualified UHF.Parts.UnifiedFrontendSolver.InfixGroup as InfixGroup
 import qualified UHF.Parts.Lexer as Lexer
-import qualified UHF.Parts.NameResolve as NameResolve
+import qualified UHF.Parts.UnifiedFrontendSolver.NameResolve as NameResolve
 import qualified UHF.Parts.OptimizeANFIR as OptimizeANFIR
 import qualified UHF.Parts.Parser as Parser
 import qualified UHF.Parts.RemovePoison as RemovePoison
 import qualified UHF.Parts.ReportHoles as ReportHoles
-import qualified UHF.Parts.SolveTypes as SolveTypes
+import qualified UHF.Parts.UnifiedFrontendSolver.SolveTypes as SolveTypes
 import qualified UHF.Parts.TSBackend as TSBackend
 import qualified UHF.Parts.ToANFIR as ToANFIR
 import qualified UHF.Parts.ToBackendIR as ToBackendIR
