+++ src/UHF/Parts/UnifiedFrontendSolver/SolveTypes/Aliases.hs
@@ -1,14 +1,14 @@
-module UHF.Parts.SolveTypes.Aliases where
+module UHF.Parts.UnifiedFrontendSolver.SolveTypes.Aliases where
 
 import UHF.Prelude
 
 import qualified UHF.Data.IR.Type as Type
 import qualified UHF.Data.IR.Type.ADT as Type.ADT
 import qualified UHF.Data.SIR as SIR
-import qualified UHF.Parts.TypeSolver.TypeWithInferVar as TypeWithInferVar
+import qualified UHF.Data.SIR.Type as SIR.Type
 import qualified UHF.Util.Arena as Arena
 
-type TypeWithInferVars = TypeWithInferVar.Type
+type TypeWithInferVars = SIR.Type.Type
 type Type = Type.Type
 
 type UntypedDIden = Maybe (SIR.DeclRef TypeWithInferVars)
